post-mr fixes
